Fix access_token parsing (e.g. now contains '_')
We were failing to parse them correctly with the hacky regex,
which lead to 401 Forbiddens. I suspect that the access_token
format changed at some point and that explains why the app
suddenly stopped working. The extraction of the param is
quite hacky because we are dealing with the VueRouter
"hash history mode", unfortunately.

Also upgraded packages while debugging.